pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tidbcloud-manager"
version = "0.1.0"
description = "Local runner for TiDB Cloud E2E exploration sessions and YAML scenario generation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.1",
]

[project.scripts]
tidbcloud-manager = "tidbcloud_manager.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
